fix: added alias to support chat
