+++ packages/examples/sdk-frontend-react/src/app/app.tsx
@@ -88,6 +88,7 @@ import SearchSpaceTest from './SpaceTest/SearchSpaceTest';
 import SearchGroupTest from './ChatTest/SearchGroupTest';
 import RejectRequestTest from './ChatTest/RejectRequestTest';
 import GetGroupMemberStatusTest from './ChatTest/GetGroupMemberStatusTest';
+import { ChatSupportTest } from './ChatSupportTest';
 
 
 window.Buffer = window.Buffer || Buffer;
@@ -559,7 +560,10 @@ export function App() {
                       <Route
                         path="ChatProfile"
                         element={<ChatProfileTest />}
-                      />
+                      /> <Route
+                      path="ChatSupport"
+                      element={<ChatSupportTest />}
+                    />
                   </Routes>
                   {/* <ChatWidgetTest/> */}
                   {/* <ChatWidgetTest /> */}
